Add typedef struct for output buffer and handle available output space
The struct contains a pointer to the output buffer and the length
of the buffer. This is necessary for handling the amount of available
space in the output buffer. The JSON_WRITE will now return 1 as long
as the JSON serialization has not been finished. Re-calling JSON_WRITE
will continue from the point, where the output buffer was filled last.