Redo how jemi processes a node pool
Rework is based on wish for serializing JSON in chunks. Instead of
recursively calling the 'emit_aux', all the 'emit_*' functions now
return a pointer to a node. The 'while' loop for processing the linked
list is moved to 'jemi_emit' and will loop until the output is full.
A check-point node is returned, so calling 'jemi_emit' again on this
will pickup from where the serialization ended. 'jemi_emit' returns
NULL when list has been fully processed.

Also, nodes now contain both their (potential) key and value, so a
reduced number of nodes is necessary for a given JSON schema.